Docs on retry settings
