+++ jms/src/main/resources/reference.conf
@@ -2,10 +2,15 @@ alpakka.jms {
   # Connection Retry Settings
   # these set the defaults for Consumer, Producer, and Browse settings
   connection-retry {
+    # Time allowed to establish and start a connection.
     connect-timeout = 10 seconds
+    # Wait time before retrying the connection the first time.
     initial-retry = 100 millis
+    # Back-off factor for subsequent retries.
     backoff-factor = 2
+    # Back-off factor for subsequent retries.
     max-backoff = 1 minute
+    # Maximum number of retries allowed.
     # "infinite", or positive integer
     max-retries = 10
   }
@@ -34,9 +39,14 @@ alpakka.jms {
   # Send Retry Settings
   # these set the defaults for Producer settings
   send-retry {
+    # Wait time before retrying the first time.
     initial-retry = 20 millis
+    # Back-off factor for subsequent retries.
     backoff-factor = 1.5
+    # Maximum back-off time allowed, after which all retries will happen after this delay.
     max-backoff = 500 millis
+    # Maximum number of retries allowed.
+    # "infinite", or positive integer
     max-retries = 10
   }
